Extract @bb/tunnel-client from the connect plugin
Move transport-generic TunnelSession machinery (headers rewrite, heartbeat,
humanized transport errors, reconnect backoff) into a shared package so the
daemon can reuse it for machine-direct tunnels without depending on the
plugin. Connect keeps pairing, credentials, shares, and status; zero behavior
change.
