+++ packages/tunnel-client/src/index.ts
@@ -1,0 +1,20 @@
+export type { TunnelClientLogger } from "./logger.js";
+export {
+  headersForLoopbackRequest,
+  type LoopbackHeaderRewrite,
+} from "./headers.js";
+export { humanizeTransportError } from "./humanize.js";
+export {
+  DEFAULT_MAX_RECONNECT_DELAY_MS,
+  DEFAULT_RECONNECT_BASE_DELAY_MS,
+  DEFAULT_STABLE_CONNECTION_MS,
+  ReconnectBackoff,
+  type ReconnectBackoffOptions,
+} from "./reconnect.js";
+export {
+  isBareBbRealtimeWs,
+  TunnelSession,
+  type ResolvedStreamOrigin,
+  type StreamOriginResult,
+  type TunnelSessionOptions,
+} from "./session.js";
